+++ nes-core/src/Sources/Parsers/Parser.cpp
@@ -34,7 +34,8 @@ void Parser::writeFieldValueToTupleBuffer(std::string inputString,
                                           uint64_t schemaFieldIndex,
                                           Runtime::MemoryLayouts::DynamicTupleBuffer& tupleBuffer,
                                           const SchemaPtr& schema,
-                                          uint64_t tupleCount) {
+                                          uint64_t tupleCount,
+                                          const Runtime::BufferManagerPtr& bufferManager) {
     auto fields = schema->fields;
     auto dataType = fields[schemaFieldIndex]->getDataType();
     auto physicalType = DefaultPhysicalTypeFactory().getPhysicalType(dataType);
@@ -110,8 +111,27 @@ void Parser::writeFieldValueToTupleBuffer(std::string inputString,
                 break;
             }
             case NES::BasicPhysicalType::TEXT: {
-                char* value = tupleBuffer[tupleCount][schemaFieldIndex].read<char*>();
-                strcpy(value, inputString.c_str());
+                NES_TRACE("Parser::writeFieldValueToTupleBuffer(): trying to write the variable length input string: "
+                          << inputString << "to tuple buffer");
+
+                auto sizeOfInputField = inputString.size();
+                auto totalSize = sizeOfInputField + sizeof(uint32_t);
+                auto childTupleBuffer = allocateVariableLengthField(bufferManager, totalSize);
+
+                NES_ASSERT(childTupleBuffer.getBufferSize() >= totalSize,
+                           "Parser::writeFieldValueToTupleBuffer(): Could not write TEXT field to tuple buffer, there was not "
+                           "sufficient space available. Required space: "
+                               << totalSize << ", available space: " << childTupleBuffer.getBufferSize());
+
+                // write out the length and the variable-sized text to the child buffer
+                (*childTupleBuffer.getBuffer<uint32_t>()) = sizeOfInputField;
+                std::memcpy(childTupleBuffer.getBuffer() + sizeof(uint32_t), inputString.c_str(), sizeOfInputField);
+
+                // attach the child buffer to the parent buffer and write the child buffer index in the
+                // schema field index of the tuple buffer
+                auto childIdx = tupleBuffer.getBuffer().storeChildBuffer(childTupleBuffer);
+                tupleBuffer[tupleCount][schemaFieldIndex].write<Runtime::TupleBuffer::NestedTupleBufferKey>(childIdx);
+
                 break;
             }
             case NES::BasicPhysicalType::BOOLEAN: {